feat: compile and run on Unity 6.5, via an EntityId compat shim
The package did not compile on Unity 6.5. GetInstanceID, InstanceIDToObject
and objectReferenceInstanceIDValue became obsolete-as-error there when the
identifier widened to 64 bits — ten call sites across six files, found when
the release-gate script picked the newest installed Editor.

All ten now go through one place, EntityIdCompat, which spells the operation
with EntityId on 6.5+ and with the int API below it. The wire contract does
not move: the field is still called instanceId and is still a JSON number,
widened from int to long so a 64-bit id survives instead of being truncated
through int — the one mistake the Unity migration guide calls out by name.

The boundary is 6.5, and getting it right needed the Editors, not a guess.
The first cut split on UNITY_6000_2_OR_NEWER, on the basis that GetEntityId
arrived in 6.2. It compiled against 6.5's references and passed, then failed
to compile on 6.3: EntityId.ToULong, FromULong and objectReferenceEntityIdValue
are 6.5 additions, and 6.3 still has the int API without deprecation. So the
split is UNITY_6000_5_OR_NEWER, and 6.0 through 6.4 stay on the int path.

Verified by running the full EditMode suite on three Editors rather than
trusting the #if: 6.0 and 6.3 exercise the int path, 6.5 the EntityId path,
163 passing on each. AnInstanceIdWins round-trips an id through the shim, so
the 6.5 run actually converts through EntityId.ToULong and back rather than
only compiling the branch.

The README claimed 2022.3 through Unity 6 while not building on 6.5; it now
says which versions the suite was verified on.

+++ jp.shiranui-isuzu.unity-mcp/Editor/Tools/InspectorTools.cs
@@ -25,7 +25,7 @@ namespace UnityMCP.Editor.Tools
             [McpArg("property_path", "Serialized property path, e.g. m_LocalPosition.x.")]
             string propertyPath,
             [McpArg("instance_id", "Target object instance id; alternative to game_object_path.")]
-            int? instanceId = null,
+            long? instanceId = null,
             [McpArg("game_object_path", "Scene path of the target GameObject, e.g. Root/Child.")]
             string gameObjectPath = null,
             [McpArg("component_type", "Component type name; omit for the GameObject itself.")]
@@ -49,7 +49,7 @@ namespace UnityMCP.Editor.Tools
             Idempotency = McpIdempotency.Safe)]
         public static JObject List(
             [McpArg("instance_id", "Target object instance id; alternative to game_object_path.")]
-            int? instanceId = null,
+            long? instanceId = null,
             [McpArg("game_object_path", "Scene path of the target GameObject, e.g. Root/Child.")]
             string gameObjectPath = null,
             [McpArg("component_type", "Component type name; omit for the GameObject itself.")]
@@ -89,7 +89,7 @@ namespace UnityMCP.Editor.Tools
             [McpArg("value", "New value; its JSON type must match the property's type.")]
             JToken value,
             [McpArg("instance_id", "Target object instance id; alternative to game_object_path.")]
-            int? instanceId = null,
+            long? instanceId = null,
             [McpArg("game_object_path", "Scene path of the target GameObject, e.g. Root/Child.")]
             string gameObjectPath = null,
             [McpArg("component_type", "Component type name; omit for the GameObject itself.")]
